feat(machine edit): Add edit page

+++ pages/api/machine-types/[id].js
@@ -2,31 +2,42 @@ import { NextApiRequest, NextApiResponse } from "next";
 
 import MachineryType from "../../../models/MachineryType";
 import connectDB from "../../../middleware/mongodb";
+import { Machine } from "@/models";
 
 /** 
 * @param {NextApiRequest} req
 * @param {NextApiResponse} res
 */
 const handler = async (req, res ) => {
+  console.log(req.method);
   switch (req.method) {
     case "PUT":
       const { name } = req.body;
-      MachineryType.findByIdAndUpdate(
+      await MachineryType.findByIdAndUpdate(
         req.query.id,
         { name: name },
         (err, result) => {
           res.json(result);
+          res.redirect("/machine-types")
         }
       );
       break;
     case "GET":
       MachineryType.findById(req.query.id, (err, result) => {
+        console.log(result);
         res.json(result);
       });
+      break;
     case "DELETE":
-      MachineryType.findByIdAndDelete(req.query.id, (err, result) => {
+      const machine = await Machine.findOne({machine_type: req.query.id})
+      if (machine) {
+        console.log("Found one");
+        break;
+      }
+      await MachineryType.findByIdAndDelete(req.query.id, (err, result) => {
         res.json(result)
       })
+      break;
     default:
       break;
   }
